+++ CHANGES.rst
@@ -18,6 +18,7 @@ Changelog
 * Rework API documentation.
 * Add unittest to cover unicode names for Cask packages. Closes #16.
 * Add unittest to cover Cask packages with multiple names. Refs #26.
+* Drop support of Python 3.3.
 
 
 `2.4.0 (2017-01-28) <https://github.com/kdeldycke/meta-package-manager/compare/v2.3.0...v2.4.0>`_